backend/requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# ─── Database ────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.3
psycopg2-binary==2.9.9

# ─── Cache ───────────────────────────────────────────────────────────────────
redis==5.1.1

# ─── AI / LLM ─────────────────────────────────────────────────────────────────
anthropic==0.40.0
langgraph==0.2.55
langchain-core==0.3.22

# ─── Gmail ───────────────────────────────────────────────────────────────────
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.151.0

# ─── Email Parsing ────────────────────────────────────────────────────────────
beautifulsoup4==4.12.3
lxml==5.3.0
html2text==2024.2.26

# ─── Data Validation ─────────────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.2.0

# ─── HTTP Client ─────────────────────────────────────────────────────────────
httpx==0.27.2

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
python-dateutil==2.9.0
pytz==2024.2
structlog==24.4.0
tenacity==9.0.0

# ─── Dev ─────────────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
